pyproject.toml
[project]
name = "knee-segment"
version = "0.1.0"
description = "Modular U-Net segmentation pipeline for knee ultrasound images"
authors = [{name = "Your Name", email = "email@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # For CUDA support, you may need to install from PyTorch index:
    # torch --index-url https://download.pytorch.org/whl/cu118
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "opencv-python>=4.8.0",
    "scikit-image>=0.21.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pillow>=10.0.0",
    "tqdm>=4.65.0",
    "tensorboard>=2.13.0",
    "pydantic>=2.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "albumentations>=1.3.0",
    "wandb>=0.16.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mlflow>=2.5.0",
    "optuna>=3.2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.4.0",
    "ipykernel>=6.25.0",
    "jupyter>=1.0.0",
]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "UP", "B", "SIM", "C4", "PTH", "RET", "ARG"]
ignore = ["E501"]
target-version = "py310"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# Development project - no build system needed 